requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
#
psycopg2-binary==2.9.9
# env config
python-decouple
python-dotenv
# unittest
pytest-django==4.9.0
pytest==8.3.3
model-bakery==1.19.5
# for mypy
mypy==1.11.2
django-stubs
djangorestframework-stubs

# pep8
flake8
autopep8

isort
